Refactor, Add comments, Fix bugs

+++ baseline_knn_chi2_webKB.py
@@ -14,9 +14,10 @@ X = data["X"]
 vectorizer = TfidfVectorizer()
 X = vectorizer.fit_transform(X.values.astype('U')).toarray()
 
-x_train = SelectKBest(chi2, k=0.1*len(X[1])).fit_transform(X, Y)
+x_train = SelectKBest(chi2, k=int(0.1*len(X[1]))).fit_transform(X, Y)
 y_train = Y
 
+# Uncomment to use test dataset:
 # data = pd.read_table("data/WebKB/webkb-test-stemmed.txt")
 # data.columns=["Y","X"]
 # classes = {'project':0, 'faculty':1, 'course':2, 'student':3}
